packages/platform-fast-server/test/version-filter/chain/uri-version-handler.test.ts
import { mock, describe, it, expect, beforeEach } from 'bun:test';
import { AbstractHandler } from '../../../adapters/version-filter/abstract-handler';
import { UriVersioningHandler } from '../../../adapters/version-filter/chain/uri-version-handler';

class AbstractHandlerImp extends AbstractHandler{}

describe('UriVersioningHandler and AbstractHandler Tests', () => {
  let uriVersioningHandler;
  let abstractHandler;
  let mockHandler;
  let mockReq;
  let mockRes;
  let mockNext;
  let version;
  let versioningOptions;
  let originalHandler;

  beforeEach(() => {
    uriVersioningHandler = new UriVersioningHandler();
    abstractHandler = new AbstractHandlerImp();

    mockHandler = {
      handle: mock(() => this)
    };

    mockReq = {}; // Adjust this as per the structure of your request object
    mockRes = {}; // Adjust this as per the structure of your response object
    mockNext = mock(() => {});
    version = '1.0';
    versioningOptions = {}; // Adjust based on your versioning options
    originalHandler = mock(() => {});
  });

  it('UriVersioningHandler should call originalHandler if version is neutral', () => {
    uriVersioningHandler.handle(mockReq, mockRes, mockNext, 'VERSION_NEUTRAL', versioningOptions, originalHandler);
    expect(originalHandler).toHaveBeenCalled();
  });

  it('UriVersioningHandler should call super.handle if version is not neutral', () => {
    uriVersioningHandler.handle(mockReq, mockRes, mockNext, version, versioningOptions, originalHandler);
    // Here you need to verify if super.handle was called. This might need a spy or mock of super.handle
    expect(originalHandler).toHaveBeenCalled();
  });

  it('AbstractHandler setNext should set the next handler', () => {
    abstractHandler.setNext(mockHandler);
    // Verify if the nextHandler property is set correctly
    expect(abstractHandler['nextHandler']).toEqual(mockHandler);
  });

  it('AbstractHandler should call next handler if exists', () => {
    abstractHandler.setNext(mockHandler);
    abstractHandler.handle(mockReq, mockRes, mockNext, version, versioningOptions, originalHandler);
    expect(mockHandler.handle).toHaveBeenCalled();
  });

  it('AbstractHandler should call originalHandler if no next handler', () => {
    abstractHandler.handle(mockReq, mockRes, mockNext, version, versioningOptions, originalHandler);
    expect(originalHandler).toHaveBeenCalled();
  });
});
